Improves hotfix and alpha release validation
Enhances validation for hotfix releases by restricting them to 'release' or 'main' branches.

Adds a check for open pull requests before allowing alpha releases to proceed, ensuring code stability.

Introduces validation that the reference is a hotfix branch when doing a hotfix release to prevent the usage of target branch as reference branch.