Refactors release tag retrieval logic
Improves the way release tags are filtered and sorted. It introduces a function to retrieve the latest version tag for a given branch (develop, release, or main).

This change simplifies the code and makes it more readable. It also fixes an issue with the release tag regex that prevented beta versions with multiple sub-versions from being matched.

The main motivation is to ensure version numbers are accurately determined for hotfix branch creation on the develop branch.